+++ zeroeq/publisher.cpp
@@ -1,10 +1,11 @@
 
-/* Copyright (c) 2014-2016, Human Brain Project
+/* Copyright (c) 2014-2017, Human Brain Project
  *                          Daniel Nachbaur <daniel.nachbaur@epfl.ch>
  *                          Stefan.Eilemann@epfl.ch
  */
 
 #include "publisher.h"
+#include "detail/application.h"
 #include "detail/broker.h"
 #include "detail/byteswap.h"
 #include "detail/constants.h"
@@ -13,55 +14,12 @@
 
 #include <servus/serializable.h>
 #include <servus/servus.h>
-#if __APPLE__
-#include <dirent.h>
-#include <mach-o/dyld.h>
-#endif
 
 #include <cstring>
 #include <map>
 
 namespace zeroeq
 {
-namespace
-{
-std::string _getApplicationName()
-{
-// http://stackoverflow.com/questions/933850
-#ifdef _MSC_VER
-    char result[MAX_PATH];
-    const std::string execPath(result,
-                               GetModuleFileName(NULL, result, MAX_PATH));
-#elif __APPLE__
-    char result[PATH_MAX + 1];
-    uint32_t size = sizeof(result);
-    if (_NSGetExecutablePath(result, &size) != 0)
-        return std::string();
-    const std::string execPath(result);
-#else
-    char result[PATH_MAX];
-    const ssize_t count = readlink("/proc/self/exe", result, PATH_MAX);
-    if (count < 0)
-    {
-        // Not all UNIX have /proc/self/exe
-        ZEROEQWARN << "Could not find absolute executable path" << std::endl;
-        return std::string();
-    }
-    const std::string execPath(result, count > 0 ? count : 0);
-#endif
-
-#ifdef _MSC_VER
-    const size_t lastSeparator = execPath.find_last_of('\\');
-#else
-    const size_t lastSeparator = execPath.find_last_of('/');
-#endif
-    if (lastSeparator == std::string::npos)
-        return execPath;
-    // lastSeparator + 1 may be at most equal to filename.size(), which is good
-    return execPath.substr(lastSeparator + 1);
-}
-}
-
 class Publisher::Impl : public detail::Sender
 {
 public:
@@ -148,7 +106,7 @@ private:
 
         _service.set(KEY_INSTANCE, detail::Sender::getUUID().getString());
         _service.set(KEY_USER, getUserName());
-        _service.set(KEY_APPLICATION, _getApplicationName());
+        _service.set(KEY_APPLICATION, detail::getApplicationName());
         if (!_session.empty())
             _service.set(KEY_SESSION, _session);